# docs/openapi-skeleton.yaml
openapi: 3.0.3
info:
  title: Tidecove
  version: 0.1.0
  description: >-
    Enterprise ai-data exchange with publish-once offers across marketplaces without mandatory central hosting.
  x-product-slug: ai-decentralized-tidecove
  x-source: ai-in-decentralized+ai/openminedOcean Protocol Business Whitepaper/
servers:
  - url: https://api.example.com
...
paths:
...
  /v1/marketplace-fronts:
    get:
      operationId: listMarketplaceFronts
      tags: [MarketplaceFronts]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Cursor" }, { $ref: "#/components/parameters/Limit" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListEnvelopeMarketplaceFront" }
    post:
      operationId: registerMarketplaceFront
      tags: [MarketplaceFronts]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarketplaceFrontCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DataEnvelopeMarketplaceFront" }
  /v1/marketplace-fronts/{frontId}:
    parameters: [{ $ref: "#/components/parameters/FrontId" }]
    get:
      operationId: getMarketplaceFront
      tags: [MarketplaceFronts]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DataEnvelopeMarketplaceFront" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/settlements:
    get:
      operationId: listSettlementStatements
      tags: [Settlements]
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Cursor" }, { $ref: "#/components/parameters/Limit" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListEnvelopeSettlementStatement" }
  /v1/settlements/{statementId}:
    parameters: [{ $ref: "#/components/parameters/StatementId" }]
    get:
      operationId: getSettlementStatement
      tags: [Settlements]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DataEnvelopeSettlementStatement" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
    FrontId:
      name: frontId
      in: path
      required: true
      schema: { type: string }
...
    StatementId:
      name: statementId
      in: path
      required: true
      schema: { type: string }
  responses:
...
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    MarketplaceFront:
      type: object
      required: [id, name, takeRateBps, status]
      properties:
        id: { type: string }
        name: { type: string }
        takeRateBps: { type: integer }
        status:
          type: string
          enum: [active, suspended]
        vertical: { type: string }
    MarketplaceFrontCreate:
      type: object
      required: [name, takeRateBps, status]
      properties:
        name: { type: string }
        takeRateBps: { type: integer }
        status:
          type: string
          enum: [active, suspended]
        vertical: { type: string }
    DataEnvelopeMarketplaceFront:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/MarketplaceFront" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    ListEnvelopeMarketplaceFront:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/MarketplaceFront" }
            nextCursor: { type: string }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
...
    SettlementStatement:
      type: object
      required: [id, participantId, periodStart, periodEnd, currency, netAmount]
      properties:
        id: { type: string }
        participantId: { type: string }
        periodStart: { type: string, format: date-time }
        periodEnd: { type: string, format: date-time }
        currency: { type: string }
        netAmount: { type: number }
...
    DataEnvelopeSettlementStatement:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/SettlementStatement" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    ListEnvelopeSettlementStatement:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/SettlementStatement" }
            nextCursor: { type: string }
        meta: { $ref: "#/components/schemas/ResponseMeta" }