# platform/api-server/src/domains/marketplace/handlers/marketplace-front.handlers.ts
/**
 * Handlers for MarketplaceFront
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: marketplace
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { MarketplaceDomainModule } from "../dependencies/marketplace-ddd.dependencies.js";
import type { GetMarketplaceFrontInput, ListMarketplaceFrontsInput, RegisterMarketplaceFrontInput, UpdateMarketplaceFrontInput } from "@tidecove/services/marketplace";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/marketplace-fronts
 * Operation: listMarketplaceFronts
 */
export async function listMarketplaceFronts(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: MarketplaceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListMarketplaceFrontsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.fronts.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/marketplace-fronts
 * Operation: registerMarketplaceFront
 */
export async function registerMarketplaceFront(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: MarketplaceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: RegisterMarketplaceFrontInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.fronts.create.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/marketplace-fronts/{frontId}
 * Operation: getMarketplaceFront
 */
export async function getMarketplaceFront(
  request: FastifyRequest<{ Params: { frontId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: MarketplaceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetMarketplaceFrontInput = {
    frontId: request.params.frontId,
  };
  const result = await deps.useCases.fronts.get.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for PATCH /v1/marketplace-fronts/{frontId}
 * Operation: updateMarketplaceFront
 */
export async function updateMarketplaceFront(
  request: FastifyRequest<{ Params: { frontId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: MarketplaceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: UpdateMarketplaceFrontInput = {
    frontId: request.params.frontId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.fronts.update.execute(input);
  return reply.code(200).send(result);
}