# platform/api-server/src/domains/settlements/handlers/settlement.handlers.ts
/**
 * Handlers for Settlement
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: settlements
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { SettlementsDomainModule } from "../dependencies/settlements-ddd.dependencies.js";
import type { GenerateSettlementStatementInput, GetSettlementStatementInput, ListSettlementStatementsInput } from "@tidecove/services/settlements";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/settlements
 * Operation: listSettlementStatements
 */
export async function listSettlementStatements(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], participantId?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: SettlementsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListSettlementStatementsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.settlements.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/settlements
 * Operation: generateSettlementStatement
 */
export async function generateSettlementStatement(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: SettlementsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GenerateSettlementStatementInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.settlements.get.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/settlements/{statementId}
 * Operation: getSettlementStatement
 */
export async function getSettlementStatement(
  request: FastifyRequest<{ Params: { statementId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: SettlementsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetSettlementStatementInput = {
    statementId: request.params.statementId,
  };
  const result = await deps.useCases.settlements.get.execute(input);
  return reply.code(200).send(result);
}